+++ crates/crb-agent/src/lib.rs
@@ -9,7 +9,7 @@ pub mod runtime;
 
 pub use address::{Address, MessageFor};
 pub use agent::{Agent, Runnable, Standalone};
-pub use context::{AgentContext, AgentSession};
+pub use context::{AgentContext, AgentSession, Context};
 pub use equip::Equip;
 pub use message::event::OnEvent;
 pub use performers::async_performer::DoAsync;