+++ crates/crb-agent/src/message/interrupt.rs
@@ -1,6 +1,6 @@
-use crate::address::Address;
-use crate::address::MessageFor;
+use crate::address::{Address, MessageFor};
 use crate::agent::Agent;
+use crate::context::Context;
 use anyhow::Result;
 use async_trait::async_trait;
 
@@ -14,7 +14,7 @@ struct Interrupt;
 
 #[async_trait]
 impl<A: Agent> MessageFor<A> for Interrupt {
-    async fn handle(self: Box<Self>, agent: &mut A, ctx: &mut A::Context) -> Result<()> {
+    async fn handle(self: Box<Self>, agent: &mut A, ctx: &mut Context<A>) -> Result<()> {
         agent.interrupt(ctx);
         Ok(())
     }